refactor(compression): extract risk-gate strategy wrappers + de-complex preview POST
